backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
pymongo==4.5.0
mongoengine==0.27.0

# Security
bcrypt==4.0.1
Werkzeug==2.3.7

# Validation and serialization
marshmallow==3.20.1
email-validator==2.0.0

numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-mock==3.11.1

# Development
autopep8==2.0.4
flake8==6.1.0

# Production server
gunicorn==21.2.0

# Logging and monitoring
structlog==23.1.0